+++ controllers/device_del_data.php
@@ -1,4 +1,8 @@
 <?php
+include('_session_use.php');
+include('_session_check.php');
+$user_id = $_SESSION['session_key'];
+
 $data = array();
 
 include("../helpers/mysqli_connect.php");
@@ -6,7 +10,7 @@ include("../helpers/mysqli_connect.php");
 if(!empty($_POST['id'])){
     try {
         //del data from the database
-        $sql = "DELETE FROM devices WHERE device_id = {$_POST['id']} ";
+        $sql = "DELETE FROM devices WHERE device_id = {$_POST['id']} AND device_user_id = $user_id ";
         $conn->query($sql);
 
         $data = array("result" => true,
